learn initial focus vector
This is a much simpler alternative to my old approach, which was to have a filler timestep at the beginning of the sequence so that the RNN got a chance to attend to something. Now, the network will always attend to some initial vector for the first timestep, producing its first output immediately. This removes the need for a bunch of weird first-timestep-clipping-voodoo.